refactor: split dialogs
